separated NetworkModule
