+++ client/app/src/main/java/io/mosip/registration/app/AppComponent.java
@@ -9,6 +9,7 @@ import dagger.Component;
 import dagger.android.AndroidInjector;
 import dagger.android.support.AndroidSupportInjectionModule;
 import io.mosip.registration.clientmanager.config.AppModule;
+import io.mosip.registration.clientmanager.config.NetworkModule;
 import io.mosip.registration.clientmanager.config.RoomModule;
 
 @Singleton
@@ -16,6 +17,7 @@ import io.mosip.registration.clientmanager.config.RoomModule;
         modules = {
                 AndroidSupportInjectionModule.class,
                 ActivityBuildersModule.class,
+                NetworkModule.class,
                 RoomModule.class,
                 AppModule.class
         }
@@ -26,6 +28,7 @@ public interface AppComponent extends AndroidInjector<BaseApplication> {
     interface Builder{
         @BindsInstance
         Builder application(Application application);
+        Builder networkModule(NetworkModule networkModule);
         Builder roomModule(RoomModule roomModule);
         Builder appModule(AppModule appModule);
         AppComponent build();
