+++ client/app/src/main/java/io/mosip/registration/app/BaseApplication.java
@@ -3,6 +3,7 @@ package io.mosip.registration.app;
 import dagger.android.AndroidInjector;
 import dagger.android.DaggerApplication;
 import io.mosip.registration.clientmanager.config.AppModule;
+import io.mosip.registration.clientmanager.config.NetworkModule;
 import io.mosip.registration.clientmanager.config.RoomModule;
 
 public class BaseApplication extends DaggerApplication {
@@ -10,7 +11,9 @@ public class BaseApplication extends DaggerApplication {
 
     @Override
     protected AndroidInjector<? extends DaggerApplication> applicationInjector() {
-        return DaggerAppComponent.builder().application(this)
+        return DaggerAppComponent.builder()
+                .application(this)
+                .networkModule(new NetworkModule(this))
                 .roomModule(new RoomModule(this))
                 .appModule(new AppModule(this))
                 .build();
