+++ client/clientmanager/src/main/java/io/mosip/registration/clientmanager/config/AppModule.java
@@ -8,22 +8,15 @@ import androidx.annotation.NonNull;
 import javax.inject.Singleton;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.gson.FieldNamingPolicy;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 
 import dagger.Module;
 import dagger.Provides;
-import io.mosip.registration.clientmanager.BuildConfig;
 import io.mosip.registration.clientmanager.service.*;
 import io.mosip.registration.clientmanager.spi.JobManagerService;
-import io.mosip.registration.clientmanager.util.LocalDateTimeDeserializer;
-import io.mosip.registration.clientmanager.util.LocalDateTimeSerializer;
 import io.mosip.registration.clientmanager.spi.AuditManagerService;
 import io.mosip.registration.clientmanager.spi.JobTransactionService;
 import io.mosip.registration.clientmanager.spi.RegistrationService;
 import io.mosip.registration.clientmanager.util.SyncRestUtil;
-import io.mosip.registration.clientmanager.interceptor.RestAuthInterceptor;
 import io.mosip.registration.clientmanager.repository.*;
 import io.mosip.registration.clientmanager.spi.MasterDataService;
 import io.mosip.registration.clientmanager.spi.PacketService;
@@ -46,12 +39,6 @@ import io.mosip.registration.packetmanager.spi.ObjectAdapterService;
 import io.mosip.registration.packetmanager.spi.PacketWriterService;
 import io.mosip.registration.packetmanager.util.PacketKeeper;
 import io.mosip.registration.packetmanager.util.PacketManagerHelper;
-import okhttp3.Cache;
-import okhttp3.OkHttpClient;
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
-
-import java.time.LocalDateTime;
 
 @Module
 public class AppModule {
@@ -146,47 +133,7 @@ public class AppModule {
                 caCertificateManagerService, languageRepository, jobManagerService);
     }
 
-    @Provides
-    @Singleton
-    Cache provideHttpCache() {
-        int cacheSize = 10 * 1024 * 1024;
-        return new Cache(application.getCacheDir(), cacheSize);
-    }
 
-    @Provides
-    @Singleton
-    Gson provideGson() {
-        GsonBuilder gsonBuilder = new GsonBuilder();
-        gsonBuilder.setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES);
-        gsonBuilder.registerTypeAdapter(LocalDateTime.class, new LocalDateTimeSerializer());
-        gsonBuilder.registerTypeAdapter(LocalDateTime.class, new LocalDateTimeDeserializer());
-        return gsonBuilder.create();
-    }
-
-    @Provides
-    @Singleton
-    OkHttpClient provideOkhttpClient(Cache cache) {
-        OkHttpClient.Builder client = new OkHttpClient.Builder();
-        client.cache(cache);
-        client.addInterceptor(new RestAuthInterceptor(appContext));
-        return client.build();
-    }
-
-    @Provides
-    @Singleton
-    Retrofit provideRetrofit(Gson gson, OkHttpClient okHttpClient) {
-        return new Retrofit.Builder()
-                .addConverterFactory(GsonConverterFactory.create(gson))
-                .baseUrl(BuildConfig.BASE_URL)
-                .client(okHttpClient)
-                .build();
-    }
-
-    @Provides
-    @Singleton
-    SyncRestService provideSyncRestService(Retrofit retrofit) {
-        return retrofit.create(SyncRestService.class);
-    }
 
     @Provides
     @Singleton
@@ -211,8 +158,8 @@ public class AppModule {
                                                    KeyStoreRepository keyStoreRepository,
                                                    GlobalParamRepository globalParamRepository,
                                                    AuditManagerService auditManagerService) {
-        return new RegistrationServiceImpl(appContext, objectMapper, packetWriterService, userInterfaceHelperService,
-                registrationRepository, masterDataService, identitySchemaRepository, clientCryptoManagerService,
+        return new RegistrationServiceImpl(appContext, packetWriterService, registrationRepository,
+                masterDataService, identitySchemaRepository, clientCryptoManagerService,
                 keyStoreRepository, globalParamRepository, auditManagerService);
     }
